requirements/base.txt
South
redis
django-celery
RapidSMS==0.15.0
-e git://github.com/caktus/rapidsms-appointments.git#egg=appointments
-e git://github.com/caktus/rapidsms-nutrition.git#egg=nutrition
-e git://github.com/caktus/rapidsms-natal-care.git#egg=natalcare
SQLAlchemy
-e git://github.com/ryanpetrello/cleaver.git#egg=cleaver
django-reversion==1.6.1
django-reversion-compare
diff-patch-match
